+++ Changelog.md
@@ -1,5 +1,14 @@
 # BetterServ Changelog
 
+## V 0.3.4
+
+- BetterFiles now hast multiselect! You can now select multiple files and folders at once and download them all at once.
+- Added a "Download" button to the file selection menu.
+- Added Upload button to the file selection menu.
+- Added recursive folder upload. You can now upload folders with all their contents using drag and drop.
+- Added a "New Folder" button to the file selection menu.
+- Added a "Delete" button to the file selection menu.
+
 ## V 0.3.3
 
 - Clicking on files now actually opens them in the browser instead of downloading them.